Refactor fixtures
